Move TestAction to Stubs
[MAILPOET-4935]
